backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.5.0
django-filter==24.3
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
redis==5.1.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==11.0.0
openpyxl==3.2.0b1
reportlab==4.2.2
setuptools>=75.3.0
wheel>=0.45.0
pip>=25.2
gunicorn 
